+++ sdk/keyvault/azure-keyvault-keys/azure/keyvault/keys/models.py
@@ -16,6 +16,7 @@ if TYPE_CHECKING:
     from typing import Any, Dict, Optional
     from datetime import datetime
     from ._shared._generated.v7_0 import models as _models
+    from .enums import KeyOperation
 
 KeyOperationResult = namedtuple("KeyOperationResult", ["id", "value"])
 
@@ -240,8 +241,14 @@ class Key(object):
     @property
     def key_type(self):
         # type: () -> str
-        """The key's type"""
+        """The key's type. See :class:`~azure.keyvault.keys.enums.KeyType` for possible values."""
         return self._key_material.kty
+
+    @property
+    def key_operations(self):
+        # type: () -> list[KeyOperation]
+        """Permitted operations. See :class:`~azure.keyvault.keys.enums.KeyOperation` for possible values."""
+        return self._key_material.key_ops
 
 
 class DeletedKey(Key):